+++ src/storage/database/pg.test.ts
@@ -1,0 +1,9 @@
+import { escapeLike } from './pg'
+
+describe('escapeLike', () => {
+  test('escapes SQL wildcard characters', () => {
+    expect(escapeLike('%_abc')).toBe('\\%\\_abc')
+    expect(escapeLike('a%b_c')).toBe('a\\%b\\_c')
+    expect(escapeLike('plain-text')).toBe('plain-text')
+  })
+})
